Moves reducer to its own file. Adds test.
